Cleanup WalletConnect context
