+++ apps/mobile-wallet/src/features/backup/backupScreens/VerifyMnemonicScreen.tsx
@@ -1,8 +1,7 @@
 import { bip39Words } from '@alephium/shared'
 import { useFocusEffect } from '@react-navigation/native'
 import { StackScreenProps } from '@react-navigation/stack'
-import LottieView from 'lottie-react-native'
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { lazy, useCallback, useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Alert } from 'react-native'
 import Animated, { FadeIn } from 'react-native-reanimated'
