+++ apps/mobile-wallet/src/features/ecosystem/dAppBrowser/dAppBrowserSlice.ts
@@ -1,0 +1,21 @@
+import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+
+const sliceName = 'dAppBrowser'
+
+const initialState = {
+  isInEcosystemInAppBrowser: false
+}
+
+const dAppBrowserSlice = createSlice({
+  name: sliceName,
+  initialState,
+  reducers: {
+    setIsInEcosystemInAppBrowser: (state, action: PayloadAction<boolean>) => {
+      state.isInEcosystemInAppBrowser = action.payload
+    }
+  }
+})
+
+export const { setIsInEcosystemInAppBrowser } = dAppBrowserSlice.actions
+
+export default dAppBrowserSlice
